requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.8.0
slowapi>=0.1.9
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
PyPDF2>=3.0.0
sentence-transformers>=3.0.0
python-multipart>=0.0.6
